Extract creation of genome table to its own mixin class.
