+++ webapp/views/mixins.py
@@ -6,8 +6,10 @@ from views.analysis_view_metadata import analysis_views_metadata
 from views.object_type_metadata import (AmrMetadata, CogMetadata, KoMetadata,
                                         OrthogroupMetadata, PfamMetadata,
                                         VfMetadata, my_locals)
-from views.utils import (format_hmm_url, format_ko_module, format_lst_to_html,
-                         format_refseqid_to_ncbi, page2title, safe_replace)
+from views.utils import (DataTableConfig, format_genome, format_hmm_url,
+                         format_ko_module, format_lst_to_html,
+                         format_refseqid_to_ncbi, get_genomes_data, page2title,
+                         safe_replace)
 
 
 class Transform():
@@ -425,3 +427,63 @@ class ComparisonViewMixin():
         # Fix the metadata property
         self.mixin._metadata_cls = getattr(self, "_metadata_cls", None)
         return super(ComparisonViewMixin, self).dispatch(request, *args, **kwargs)
+
+
+class GenomesTableMixin():
+
+    def get_genomes_table(self):
+        genomes_data = get_genomes_data(self.db)
+
+        filenames_tax_id = self.db.get_filenames_to_taxon_id()
+        filenames_tax_id_db = pd.DataFrame.from_dict(list(filenames_tax_id.items()))
+        filenames_tax_id_db.columns = ['filename', 'taxon_id']
+        filenames_tax_id_db.index = list(filenames_tax_id_db['taxon_id'])
+        filenames_list = list(filenames_tax_id_db["filename"])
+
+        path_template = settings.BLAST_DB_PATH + "/{ext}/{filename}.{ext}"
+        link_template = '<a href="{}"> .{{ext}} </a>'.format(path_template)
+        for ext in ["faa", "fna", "ffn", "gbk"]:
+            filenames_tax_id_db[f'path_to_{ext}'] = [
+                link_template.format(filename=filename, ext=ext)
+                for filename in filenames_list]
+
+        filenames_tax_id_db = filenames_tax_id_db[
+            ["path_to_faa", "path_to_fna", "path_to_ffn", "path_to_gbk"]]
+        genomes_data = genomes_data.join(filenames_tax_id_db, on="taxon_id")
+
+        genomes_data["accession"] = genomes_data[["id", "description"]].apply(
+            format_genome, axis=1)
+
+        data_table_header = [
+            "Name",
+            "GC %",
+            "N proteins",
+            "N contigs",
+            "Size (Mbp)",
+            "Coding %",
+            "N plasmid contigs",
+            "faa seq",
+            "fna seq",
+            "ffn seq",
+            "gbk file"
+        ]
+
+        table_data_accessors = [
+            "accession",
+            "gc",
+            "n_prot",
+            "n_contigs",
+            "length",
+            "coding_density",
+            "has_plasmid",
+            "path_to_faa",
+            "path_to_fna",
+            "path_to_ffn",
+            "path_to_gbk"]
+
+        table_data = genomes_data[table_data_accessors]
+
+        return {"table_data": table_data,
+                "table_headers": data_table_header,
+                "data_table_config": DataTableConfig(),
+                "table_data_accessors": table_data_accessors}
